src/TicketMonarch/rl_captcha/requirements.txt
torch>=2.0
gymnasium>=0.29
xgboost>=2.0
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
mysql-connector-python>=9.0
python-dotenv>=1.0
